Moved GUI into seperate files
