+++ gui.h
@@ -1,0 +1,11 @@
+#include <stdio.h>
+#include <stdlib.h>
+#include <windows.h>
+#include <conio.h>
+
+int getWindowWidth();
+
+int drawTitle();
+
+void drawMenu();
+
