+++ quiz.c
@@ -430,106 +430,5 @@ size_t getRandomQuestionId(size_t blacklist[15], size_t curId, size_t lineCount)
     }
 }
 
-/*
-*   getWindowWidth
-*       calculates the number of columns in the console window
-*
-*   return value:
-*       returns number of columns in the console window
-*
-*/
-int getWindowWidth(){
-    CONSOLE_SCREEN_BUFFER_INFO csbi;
-
-    if(!GetConsoleScreenBufferInfo(GetStdHandle(STD_OUTPUT_HANDLE), &csbi)){
-        return -1;
-    }
-
-    return csbi.srWindow.Right - csbi.srWindow.Left + 1;
-}
-
-/*
-*   printTitle
-*       prints a centred title from the title.txt file
-*
-*   return value:
-*       0 - successful execution
-*       -1 - failed execution
-*/
-int printTitle(){
-    int windowWidth = getWindowWidth();
-
-    if(windowWidth == -1){
-        printf("Failed to get window width");
-        return -1;
-    }
-
-    FILE* titleFile = fopen("title.txt", "r");
-    if(titleFile == NULL){
-        printf("Failed to open file");
-        return -1;
-    }
-
-    printf("\n");
-    char line[1024];
-    while(fgets(line, sizeof(line), titleFile)){
-        strTrimNewline(line);
-
-        int padding = (windowWidth - strlen(line)) / 2;
-
-        for(size_t i = 0; i < padding; ++i){
-            printf(" ");
-        }
-
-        printf("%s\n", line);
-    }
-
-    fclose(titleFile);
-
-    return 0;
-}
-
-void drawMenu(){
-    const char* menuItems[4] = {
-        "Graj",
-        "Dodaj pytanie",
-        "O grze",
-        "Wyjscie"
-    };
-
-    int padding = 0;
-    int windowWidth = getWindowWidth();
-
-    printf("\n\n");
-    for(size_t i = 0; i < 4; ++i){
-        padding = (windowWidth - strlen(menuItems[i])) / 2;
-
-        for(size_t i = 0; i < padding; ++i){
-            printf(" ");
-        }
-
-        printf("%d. %s\n", (i+1), menuItems[i]);
-    }
-
-    int selectedOption = 0;
-    scanf("%d", &selectedOption);
-
-    switch(selectedOption){
-    case 1:
-        printf("placeholder");
-        break;
-    case 2:
-        printf("placeholder");
-        break;
-    case 3:
-        printf("placeholder");
-        break;
-    case 4:
-        printf("placeholder");
-        break;
-    default:
-        printf("You have to provide a valid number");
-    }
-}
